+++ lib/src/tree/tree_items/leaf_node/leaf_node_builder.dart
@@ -206,7 +206,7 @@ class _LeafNodeBuilderState extends State<LeafNodeBuilder> {
               ? null
               : () => nodeConfig.onSecondaryTapCancel?.call(context),
           onHover: (bool isHovered) =>
-              nodeConfig.onHover?.call(isHovered, context),
+              nodeConfig.onHoverInkWell?.call(isHovered, context),
           mouseCursor: nodeConfig.mouseCursor,
           hoverDuration: nodeConfig.hoverDuration,
           hoverColor: nodeConfig.hoverColor,